requirements.txt
numpy
pandas
scikit-learn
joblib
tensorflow
matplotlib
pillow
